requirements.txt
openai
python-dotenv
flask
flask-sqlalchemy
